FIX: Remove code ambiguity to avoid potential division by zero crash FIX: Industry density settings not calibrated properly on startup, resulting in incorrect numbers of industry chains spawning.

+++ bauer/fabrikbauer.cc
@@ -978,7 +978,7 @@ next_ware_check:
 				if(do_not_add_beyond_target_density && !fab->is_electricity_producer())
 				{
 					//Make sure that industries are not added beyond target density.
-					if((double)(1.0 / fab->get_gewichtung()) > (welt->get_target_industry_density() - welt->get_actual_industry_density()))
+					if((double)(1.0 / (double)fab->get_gewichtung()) > (welt->get_target_industry_density() - welt->get_actual_industry_density()))
 					{
 						continue;
 					}
